refactor: redefine StoppingCriterion and adapt previous classes
We also add documentation.
