+++ src/main/java/numericalLibrary/optimization/stoppingCriteria/IterationThresholdStoppingCriterion.java
@@ -1,16 +1,23 @@
 package numericalLibrary.optimization.stoppingCriteria;
 
 
-import numericalLibrary.optimization.IterativeOptimizationAlgorithm;
+import numericalLibrary.optimization.algorithms.IterativeOptimizationAlgorithm;
 
 
 
+/**
+ * Implements the stopping criterion to stop iterating if an iteration threshold is reached.
+ */
 public class IterationThresholdStoppingCriterion
     implements StoppingCriterion
 {
     ////////////////////////////////////////////////////////////////
     // PRIVATE VARIABLES
     ////////////////////////////////////////////////////////////////
+    
+    /**
+     * Iteration threshold that defines when to stop iterating.
+     */
     private final int iterationThreshold;
     
     
@@ -19,6 +26,11 @@ public class IterationThresholdStoppingCriterion
     // PUBLIC CONSTRUCTORS
     ////////////////////////////////////////////////////////////////
     
+    /**
+     * Constructs a {@link IterationThresholdStoppingCriterion}.
+     * 
+     * @param theIterationThreshold     iteration threshold that defines when to stop iterating.
+     */
     public IterationThresholdStoppingCriterion( int theIterationThreshold )
     {
         this.iterationThreshold = theIterationThreshold;
@@ -30,9 +42,21 @@ public class IterationThresholdStoppingCriterion
     // PUBLIC METHODS
     ////////////////////////////////////////////////////////////////
     
+    /**
+     * {@inheritDoc}
+     */
+    public void initialize()
+    {
+        // There is nothing to do.
+    }
+    
+    
+    /**
+     * {@inheritDoc}
+     */
     public boolean isFinished( IterativeOptimizationAlgorithm<?> iterativeAlgorithm )
     {
-        return ( iterativeAlgorithm.getIterationLast() >= this.iterationThreshold );
+        return ( iterativeAlgorithm.getIteration() >= this.iterationThreshold );
     }
     
 }
