+++ src/main/java/numericalLibrary/optimization/stoppingCriteria/MaximumIterationsWithoutImprovementStoppingCriterion.java
@@ -1,31 +1,49 @@
 package numericalLibrary.optimization.stoppingCriteria;
 
 
-import numericalLibrary.optimization.IterativeOptimizationAlgorithm;
+import numericalLibrary.optimization.algorithms.IterativeOptimizationAlgorithm;
 
 
 
+/**
+ * Implements the stopping criterion to stop iterating if the best cost obtained so far has not been improved after a number of iterations.
+ */
 public class MaximumIterationsWithoutImprovementStoppingCriterion
     implements StoppingCriterion
 {
     ////////////////////////////////////////////////////////////////
     // PRIVATE VARIABLES
     ////////////////////////////////////////////////////////////////
+    
+    /**
+     * The algorithm is considered finished if the number of iterations without improving the best cost obtained so far surpasses this number.
+     */
     private final int maxIterationsWithoutImprovement;
     
+    /**
+     * Best cost obtained so far.
+     */
+    private double costBest;
+    
+    /**
+     * Current number of iterations without improving the best cost obtained so far.
+     */
     private int iterationsWithoutImprovement;
     
-    private double errorBestLast;
     
     
     ////////////////////////////////////////////////////////////////
     // PUBLIC CONSTRUCTORS
     ////////////////////////////////////////////////////////////////
     
+    /**
+     * Constructs a {@link MaximumIterationsWithoutImprovementStoppingCriterion}.
+     * 
+     * @param maximumIterationsWithoutImprovement   maximum number of iterations without improving the best cost obtained so far.
+     */
     public MaximumIterationsWithoutImprovementStoppingCriterion( int maximumIterationsWithoutImprovement )
     {
         this.maxIterationsWithoutImprovement = maximumIterationsWithoutImprovement;
-        this.errorBestLast = Double.MAX_VALUE;
     }
     
     
@@ -34,11 +52,23 @@ public class MaximumIterationsWithoutImprovementStoppingCriterion
     // PUBLIC METHODS
     ////////////////////////////////////////////////////////////////
     
+    /**
+     * {@inheritDoc}
+     */
+    public void initialize()
+    {
+        this.costBest = Double.MAX_VALUE;
+    }
+    
+    
+    /**
+     * {@inheritDoc}
+     */
     public boolean isFinished( IterativeOptimizationAlgorithm<?> iterativeAlgorithm )
     {
-        double errorBest = iterativeAlgorithm.getCostBest();
-        if( errorBest < this.errorBestLast ) {
-            this.errorBestLast = errorBest;
+        double cost = iterativeAlgorithm.getCost();
+        if( cost < this.costBest ) {
+            this.costBest = cost;
             this.iterationsWithoutImprovement = 0;
         } else {
             this.iterationsWithoutImprovement++;
