+++ src/main/java/numericalLibrary/optimization/stoppingCriteria/OrOperatorOnStoppingCriteria.java
@@ -1,17 +1,30 @@
 package numericalLibrary.optimization.stoppingCriteria;
 
 
-import numericalLibrary.optimization.IterativeOptimizationAlgorithm;
+import numericalLibrary.optimization.algorithms.IterativeOptimizationAlgorithm;
 
 
 
+/**
+ * Defines the OR operator for {@link StoppingCriterion}s.
+ * <p>
+ * Stop iterating if some {@link #isFinished(IterativeOptimizationAlgorithm)} returns true.
+ */
 public class OrOperatorOnStoppingCriteria
     implements StoppingCriterion
 {
     ////////////////////////////////////////////////////////////////
     // PRIVATE VARIABLES
     ////////////////////////////////////////////////////////////////
+    
+    /**
+     * First {@link StoppingCriterion} to check.
+     */
     private final StoppingCriterion firstStoppingCriterion;
+    
+    /**
+     * Second {@link StoppingCriterion} to check.
+     */
     private final StoppingCriterion secondStoppinCriterion;
     
     
@@ -20,6 +33,12 @@ public class OrOperatorOnStoppingCriteria
     // PUBLIC CONSTRUCTORS
     ////////////////////////////////////////////////////////////////
     
+    /**
+     * Constructs a {@link OrOperatorOnStoppingCriteria}.
+     * 
+     * @param first     first {@link StoppingCriterion} to check for {@link #isFinished(IterativeOptimizationAlgorithm)}.
+     * @param second    second {@link StoppingCriterion} to check for {@link #isFinished(IterativeOptimizationAlgorithm)}.
+     */
     public OrOperatorOnStoppingCriteria( StoppingCriterion first , StoppingCriterion second )
     {
         this.firstStoppingCriterion = first;
@@ -32,6 +51,19 @@ public class OrOperatorOnStoppingCriteria
     // PUBLIC METHODS
     ////////////////////////////////////////////////////////////////
     
+    /**
+     * {@inheritDoc}
+     */
+    public void initialize()
+    {
+        this.firstStoppingCriterion.initialize();
+        this.secondStoppinCriterion.initialize();
+    }
+    
+    
+    /**
+     * {@inheritDoc}
+     */
     public boolean isFinished( IterativeOptimizationAlgorithm<?> iterativeAlgorithm )
     {
         boolean isFirstFinished = this.firstStoppingCriterion.isFinished( iterativeAlgorithm );
