+++ src/main/java/numericalLibrary/optimization/stoppingCriteria/StoppingCriterion.java
@@ -1,14 +1,15 @@
 package numericalLibrary.optimization.stoppingCriteria;
 
 
-import numericalLibrary.optimization.IterativeOptimizationAlgorithm;
+import numericalLibrary.optimization.algorithms.IterativeOptimizationAlgorithm;
 
 
 
 /**
- * Represents a stopping criterion for an iterative optimization algorithm.
+ * Represents a stopping criterion for an {@link IterativeOptimizationAlgorithm}.
  * <p>
  * Stopping criteria are also known as stopping rule, stopping condition, or termination condition.
+ * They define when to stop iterating in an {@link IterativeOptimizationAlgorithm}.
  */
 public interface StoppingCriterion
 {
@@ -16,6 +17,17 @@ public interface StoppingCriterion
     // PUBLIC ABSTRACT METHODS
     ////////////////////////////////////////////////////////////////
     
+    /**
+     * Initializes the {@link StoppingCriterion}.
+     */
+    public void initialize();
+    
+    
+    /**
+     * Returns true if the {@link IterativeOptimizationAlgorithm} has finished; false otherwise.
+     * 
+     * @return  true if the {@link IterativeOptimizationAlgorithm} has finished; false otherwise.
+     */
     public boolean isFinished( IterativeOptimizationAlgorithm<?> iterativeAlgorithm );
     
 }
